+++ source/extensions/filters/network/sni_cluster/config.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "envoy/server/filter_config.h"
+#include "envoy/config/filter/network/sni_cluster/v2/sni_cluster.pb.h"
 
+#include "extensions/filters/network/common/factory_base.h"
 #include "extensions/filters/network/well_known_names.h"
 
 namespace Envoy {
@@ -13,16 +14,15 @@ namespace SniCluster {
  * Config registration for the sni_cluster filter. @see NamedNetworkFilterConfigFactory.
  */
 class SniClusterNetworkFilterConfigFactory
-    : public Server::Configuration::NamedNetworkFilterConfigFactory {
+    : public Common::FactoryBase<envoy::config::filter::network::sni_cluster::v2::SniCluster> {
+
 public:
-  // NamedNetworkFilterConfigFactory
-  Network::FilterFactoryCb createFilterFactory(const Json::Object&,
-                                               Server::Configuration::FactoryContext&) override;
-  Network::FilterFactoryCb
-  createFilterFactoryFromProto(const Protobuf::Message&,
-                               Server::Configuration::FactoryContext&) override;
-  ProtobufTypes::MessagePtr createEmptyConfigProto() override;
-  std::string name() override { return NetworkFilterNames::get().SniCluster; }
+  SniClusterNetworkFilterConfigFactory() : FactoryBase(NetworkFilterNames::get().SniCluster) {}
+
+private:
+  Network::FilterFactoryCb createFilterFactoryFromProtoTyped(
+      const envoy::config::filter::network::sni_cluster::v2::SniCluster& proto_config,
+      Server::Configuration::FactoryContext& context) override;
 };
 
 } // namespace SniCluster